requirements.txt
Django==3.2
djangorestframework==3.12.4
jsonschema==3.2.0
django-testing-utils~=0.3.0
django-filter~=2.4.0
